Changed code to use multi-threaded libraries both release and debug respectively.

+++ src/WebSocketProtocol.h
@@ -142,13 +142,13 @@ private:
             }
 
             src += payLength + MESSAGE_HEADER;
-            length -= payLength + MESSAGE_HEADER;
+            length -= (unsigned int)(payLength + MESSAGE_HEADER);
             wState->state.spillLength = 0;
             return false;
         } else {
             wState->state.spillLength = 0;
             wState->state.wantsHead = false;
-            wState->remainingBytes = payLength - length + MESSAGE_HEADER;
+            wState->remainingBytes = (unsigned int)(payLength - length + MESSAGE_HEADER);
             bool fin = isFin(src);
             if (isServer) {
                 memcpy(wState->mask, src + MESSAGE_HEADER - 4, 4);
